requirements.txt
sphinx
django
urwid
gitpython
sphinxcontrib-bibtex
sphinx-rtd-theme
jinja2